Sponsorships Section
Making a bunch of copies of the "instagram" post in preparation for when I actually clear up a few questions I have. Also fixed some stuff with the Netflix section, such as making it scroll more on smaller screens and making the section exclude the scrollbar when calculating width. Also made the prizes section scroll automatically.